Change namespace of document from Common to MarketDocument
