+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Application/ChangeOfCharges/Repositories/IMarketParticipantRepository.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using GreenEnergyHub.Charges.Domain.Common;
+using GreenEnergyHub.Charges.Domain.MarketDocument;
 
 namespace GreenEnergyHub.Charges.Application.ChangeOfCharges.Repositories
 {
