+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Domain/MarketDocument/Document.cs
@@ -13,9 +13,10 @@
 // limitations under the License.
 
 using NodaTime;
+
 #pragma warning disable 8618
 
-namespace GreenEnergyHub.Charges.Domain.Common
+namespace GreenEnergyHub.Charges.Domain.MarketDocument
 {
     // Non-nullable member is uninitialized is ignored
     // Only properties which is allowed to be null is nullable
