+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Infrastructure/Mapping/MarketParticipantMapper.cs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 using System;
-using GreenEnergyHub.Charges.Domain.Common;
+using GreenEnergyHub.Charges.Domain.MarketDocument;
 
 namespace GreenEnergyHub.Charges.Infrastructure.Mapping
 {
